Atualizar Insert OPERACOES.sql Criar rota de Login Criar rota de Lista Negra Criar rota de pegar dados do cpf

+++ src/main/java/app/models/dtos/getdadoscpf/GetDadosCPFMovimento.java
@@ -12,7 +12,7 @@ public class GetDadosCPFMovimento {
     private final Integer id;
     private final String cpf;
     private final Integer idFonte;
-    private final Integer numeroContrato;
+    private final Long numeroContrato;
     private final Date dataVencimentoParcela;
     private final Integer quantidadeParcelasVencer;
     private final Integer quantidadeParcelasPagar;
