+++ src/main/java/app/models/repository/PagamentoRepository.java
@@ -11,7 +11,7 @@ import static java.sql.DriverManager.getConnection;
 
 public class PagamentoRepository {
 
-    public List<Pagamento> selectMovimentosPorCpfENumeroContrato(String cpf, Integer contrato, Date dataVencimentoParcela) throws SQLException {
+    public List<Pagamento> selectMovimentosPorCpfENumeroContrato(String cpf, Long contrato, Date dataVencimentoParcela) throws SQLException {
         try (Connection con = getConnection(ORACLE_URL, ORACLE_USER, ORACLE_USER_PASSWORD)) {
             ResultSet resultSet;
             final String selectString = "SELECT * FROM PAGAMENTOS WHERE PGT_DOC_CLI = ? AND PGT_NUM_UNC = ? AND PGT_DAT_VCT = ?";
@@ -19,7 +19,7 @@ public class PagamentoRepository {
 
             try (PreparedStatement selectStatement = con.prepareStatement(selectString)) {
                 selectStatement.setString(1, cpf);
-                selectStatement.setInt(2, contrato);
+                selectStatement.setLong(2, contrato);
                 selectStatement.setDate(3, dataVencimentoParcela);
                 resultSet = selectStatement.executeQuery();
                 while(resultSet.next()) {
@@ -27,7 +27,7 @@ public class PagamentoRepository {
                             resultSet.getInt(1),
                             resultSet.getString(2),
                             resultSet.getInt(3),
-                            resultSet.getInt(4),
+                            resultSet.getLong(4),
                             resultSet.getDate(5),
                             resultSet.getDate(6),
                             resultSet.getBigDecimal(7),
